% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics1.R
\name{metrics1}
\alias{metrics1}
\title{Title}
\usage{
metrics1(label = NULL, pred_p = NULL, iteration = 1000, ...)
}
\arguments{
\item{label}{true label}

\item{pred_p}{predicted prob}

\item{iteration}{number of bootstrap iteration, default=1000}

\item{...}{redundant para if any}
}
\value{
point and 95\% CI dataframe
}
\description{
Title
}
\examples{
label <-  rbinom(100,size = 1,prob = 0.5)
pred_p  <- runif(100,0,1)
library(dplyr)
library(purrr)
library(pROC)
tmp <- metrics1(label,pred_p,iteration = 100)
}
