% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics1.R
\name{one_metrics}
\alias{one_metrics}
\title{Title}
\usage{
one_metrics(label = NULL, pred_p = NULL, ..., i = NULL)
}
\arguments{
\item{label}{true label}

\item{pred_p}{predicted value}

\item{...}{redundant}

\item{i}{iteration}
}
\value{
one bootstrapped point estimates of metrics: dataframe
}
\description{
Title
}
\examples{
label <-  rbinom(100,size = 1,prob = 0.5)
pred_p  <- runif(100,0,1)
library(dplyr)
library(pROC)
tmp <- one_metrics(label,pred_p)
}
