% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics1.R
\name{point_metrics}
\alias{point_metrics}
\title{Title}
\usage{
point_metrics(label = NULL, pred_p = NULL, ..., i = NULL)
}
\arguments{
\item{label}{true label}

\item{pred_p}{predicted prob}

\item{...}{XX redundant para}

\item{i}{iteration}
}
\value{
mymetrics : dataframe
}
\description{
Title
}
\examples{
label <-  rbinom(100,size = 1,prob = 0.5)
pred_p  <- runif(100,0,1)
library(dplyr)
library(pROC)
tmp <- point_metrics(label,pred_p,iteration = 100)
}
